api-server/requirements.txt
flask
paho-mqtt
kafka-python
influxdb-client
